feat: 게시글 조회수 타입을 long으로 수정

+++ src/main/java/com/stocat/amumal/post/domain/Post.java
@@ -46,7 +46,7 @@ public class Post extends BaseEntity {
   private int likeCount = 0;
 
   @Column(nullable = false)
-  private int viewCount = 0;
+  private long viewCount = 0;
 
   public static Post of(User user, String title, String content, String imageUrl) {
     Post post = new Post();
